stratiform: inline TelX/TelsX type definitions into Tel/Tels objects
`TelElement`, `TelPointer`, `TelValidator`, `TelTypeAssignment` had
been kept at the top level with type-alias + companion-singleton
shims inside `object Tel` (`type Element = TelElement`, `val Element:
TelElement.type = TelElement`, etc.). Inline their definitions
directly into `object Tel` as `Tel.Element`, `Tel.Pointer`,
`Tel.Validator`, `Tel.Type`; remove the top-level shadows; delete
the four source files.

Same for `TelsAxiom`, `TelsDecoder`, `TelsLayers`, `TelsReconstructor`
— inlined into `object Tels` as `Tels.Axiom`, `Tels.Decoder`,
`Tels.Layers`, `Tels.Reconstructor`; the four source files deleted.

`TelError` is intentionally kept at the top level. References in
`Bintel.scala`, `SchemaSignature.scala`, `soundness_stratiform_core`
and the test file are rebased onto the dotted names; the old
top-level names no longer exist.
