--- lib/stratiform/src/core/stratiform.TelsDecoder.scala
@@ -1,78 +1,0 @@
-                                                                                                  /*
-┏━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━┓
-┃                                                                                                  ┃
-┃                                                   ╭───╮                                          ┃
-┃                                                   │   │                                          ┃
-┃                                                   │   │                                          ┃
-┃   ╭───────╮╭─────────╮╭───╮ ╭───╮╭───╮╌────╮╭────╌┤   │╭───╮╌────╮╭────────╮╭───────╮╭───────╮   ┃
-┃   │   ╭───╯│   ╭─╮   ││   │ │   ││   ╭─╮   ││   ╭─╮   ││   ╭─╮   ││   ╭─╮  ││   ╭───╯│   ╭───╯   ┃
-┃   │   ╰───╮│   │ │   ││   │ │   ││   │ │   ││   │ │   ││   │ │   ││   ╰─╯  ││   ╰───╮│   ╰───╮   ┃
-┃   ╰───╮   ││   │ │   ││   │ │   ││   │ │   ││   │ │   ││   │ │   ││   ╭────╯╰───╮   │╰───╮   │   ┃
-┃   ╭───╯   ││   ╰─╯   ││   ╰─╯   ││   │ │   ││   ╰─╯   ││   │ │   ││   ╰────╮╭───╯   │╭───╯   │   ┃
-┃   ╰───────╯╰─────────╯╰────╌╰───╯╰───╯ ╰───╯╰────╌╰───╯╰───╯ ╰───╯╰────────╯╰───────╯╰───────╯   ┃
-┃                                                                                                  ┃
-┃    Soundness, version 0.54.0.                                                                    ┃
-┃    © Copyright 2021-25 Jon Pretty, Propensive OÜ.                                                ┃
-┃                                                                                                  ┃
-┃    The primary distribution site is:                                                             ┃
-┃                                                                                                  ┃
-┃        https://soundness.dev/                                                                    ┃
-┃                                                                                                  ┃
-┃    Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file     ┃
-┃    except in compliance with the License. You may obtain a copy of the License at                ┃
-┃                                                                                                  ┃
-┃        https://www.apache.org/licenses/LICENSE-2.0                                               ┃
-┃                                                                                                  ┃
-┃    Unless required by applicable law or agreed to in writing,  software distributed under the    ┃
-┃    License is distributed on an "AS IS" BASIS,  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND,    ┃
-┃    either express or implied. See the License for the specific language governing permissions    ┃
-┃    and limitations under the License.                                                            ┃
-┃                                                                                                  ┃
-┗━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━┛
-                                                                                                  */
-package stratiform
-
-import anticipation.*
-import contingency.*
-import distillate.*
-import prepositional.*
-import vacuous.*
-
-// Bridges the schema-aware semantic model into the existing
-// presentation-model-driven Tel.as[T] decoder. The fluent path for a
-// schema-validated decode is:
-//
-//   given Tels = mySchema
-//   val person: Person = tel.validate.as[Person]
-//
-// `validate` (alias `validated`) runs the §20.2 type-assignment +
-// validator pipeline against the in-scope schema, returning the same
-// Tel value if it conforms or raising TelError on the first violation
-// it encounters. The downstream `.as[Person]` then uses the existing
-// Wisteria-derived decoder, which is structurally correct because
-// type-assignment has already verified the document's shape against
-// the schema.
-
-object TelsDecoder:
-
-  extension (tel: Tel)
-    // Validate `tel` against the schema in scope and return it for
-    // chaining. Raises a TelError on the first E2xx/E3xx violation;
-    // returns the same `tel` value unchanged on success.
-    def validate(using schema: Tels): Tel raises TelError =
-      Tel.Type.assign(tel, schema)
-      tel
-
-    // Same as `validate` but also applies the registry's validators.
-    def validate(using schema: Tels, validators: Tel.Validator.Registry)
-    :     Tel raises TelError =
-      Tel.Type.assign(tel, schema, validators)
-      tel
-
-    // Convenience: validate-then-decode in a single call. Mirrors the
-    // jacinta `.as[T](using JsonSchema)` ergonomic but routed through
-    // type assignment first.
-    inline def asValidated[value: Decodable in Tel](using schema: Tels)
-    :     value raises TelError =
-      Tel.Type.assign(tel, schema)
-      tel.as[value]
